read file with char[], not as String
Read file as String is not working when reading file as a incorrect
encoding.(like reading UTF-16LE as UTF-8 blocks the readLine() forever)
So, try to read as char[] no matter if file is large or it should be
read in line.
TODO : Make LargeFileHandlingRule just a file reader class to avoid
duplicate code between reading paged/non-paged file.