line separator bug fixed
eliminate last line separator and made sure put it back when writing it
back.
In case of System line separator is longer than a char, chop off
"lastLittlePortion" that includes (possibly part of)last line separator
and put it to buffer.
Note that it makes number of char in page little bit less than user's
setting, and leftOver stores not only (portion of)last line, but also
second to last line.