to avoid deadlock in case of error, use timeout poll & add Page.ERR
Since EDT is waiting forever until worker puts a Page object, if an
Exception occurred in worker(so unable to put another value), EDT is in
deadlock.
to avoid this :

1. use poll not take for fileContentQueue. Wait for 0.5 seconds and
re-queue the task to EDT and return if Page was unavailable, allowing
EDT to do other tasks(e.g. SwingDialogs.error that queued from worker to
denote I/O Exception) that queued while waiting.
2. add Page.ERR to indicate error status or worker thread.
If worker thread catches Exception while reading or waiting for
fileContentQueue or EDT, it puts Page.ERR to fileContentQueue instead of
Page.EOF to indicate there was a Exception.
Also, add isFailed() to SelectedFileHandler so EDT could know if reading
is failed in non-paged file(paged file reading task can also use this
too).